Add jsbin to SnappySnippet
I've (@remy) have been wanting to be able to grab a blob of HTML and all it's related CSS for years now, and so pleased that someone has built this (thank you @kdzwinel!). But realised that jsbin's posting API wasn't documented (and actually had a bug when wanting to post *and* show specific panels), so I've patched the plugin to add jsbin support. Hope you like!

Cheers,

– Remy